% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_boxplot.R
\name{boxplot_diff}
\alias{boxplot_diff}
\title{multiple-panels box plot}
\usage{
boxplot_diff(score_matrix, groupinfo, paired = FALSE, fontsize = NULL)
}
\arguments{
\item{score_matrix}{a data matrix with rows labeling features immune cell type,and columns labeling samples}

\item{groupinfo}{a matrix with two columns:'sample' column labeling sampleid,'groups' column labeling group information.
when @param paired was TRUE, @param groupinfo need to provide three columns,the added column 'patient' labeling the paired relation of samples}

\item{paired}{a logical indicating whether you want a paired wilcoxon test, which is used to compare the means of two samples when each observation in one sample can be paired with an observation in the other sample,default is "FALSE"}

\item{fontsize}{a numeric value change the font size, default is NULL(function can provide the font size automatically)}
}
\description{
specify cell type variables for faceting the plot into multiple panels
}
\examples{
p <- boxplot_diff(score_matrix = cell_enrich,groupinfo = demo_groupinfo,paired = F)
print(p)
#paired sample
p <- boxplot_diff(score_matrix = cell_enrich,groupinfo = simulated_groupinfo_paired,paired = T)
print(p)
}
