% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circleTile.R
\name{circleTile}
\alias{circleTile}
\title{Add circles to a tile plot}
\usage{
circleTile(...)
}
\arguments{
\item{\dots}{Any number of arguments given below.  Must include exactly one
horizontal dimension (\code{x} or \code{top}), exactly one vertical
dimension (\code{y} or \code{right}), and should include a vector of radii
\code{r}.  All inputs should be identified by appropriate tags; i.e., use
\code{circleTile(x=myxvar, y=myyvar, r=myradii)}, \emph{not}
\code{circleTile(myxvar,myyvar,myradii)}}
}
\value{
A \code{circleTile} object, used only as an input to
\code{\link{tile}}.
}
\description{
Initializes a trace which plots circles to a tile graphic.  Useful for
making bubbleplots with known circle areas.
}
\details{
This function does no plotting; instead, it creates a \code{circleTile}
object, or \dfn{trace} of plotting data, to be drawn on one or more plots in
a tiled arrangement of plots.  To complete the drawing include the object as
an input to \code{\link{tile}}.  From \code{tile}, it is possible to set
further options including plot and axis titles, axis ranges and labels,
logged axes, and annotations to the plot.

\code{circleTile} simply adds a circle at a specific location or locations
to a \code{\link{tile}} plot.  It serves as the \pkg{tile} equivalent to the
grid graphics function \code{circleGrob}, and is primarily useful for
annotating plots which also use other traces or for constructing bubble
plots.  Note that \code{circleTile} traces must be created in advance and
included in the call to \code{tile}, rather than added afterward.

To plot symbols to a location outside the plotting area, users might try
combining \code{clip="off"}, with coordinates placing the symbols in the
desired spot.
}
\section{circleTile-specific parameters}{


A call to \code{pointsTile} \strong{must} provide an orthogonal pair of the
following inputs:

\describe{ \item{list("x")}{coordinate vector of circles to plot, attached
to the \var{x} axis.} \item{list("y")}{coordinate vector of circles to plot,
attached to the \var{y} axis.} \item{list("top")}{coordinate vector of
circles to plot, attached to the \var{top} axis.}
\item{list("right")}{coordinate vector of circles to plot, attached to the
\var{right} axis.} \item{list("r")}{radius vector of plotted circles.} }

Users will often wish to provide the following input:

\describe{ \item{list("plot")}{scalar or vector, the plot(s) in which this
trace will be drawn; defaulting to the first plot.  Plots are numbered
consecutively from the top left, row-by-row.  Thus in a 2 x 3 tiling, the
first plot in the second row is plot number 4.} }

In addition to these \code{circleTile}-specific parameters, users may
provide any of the generic tile parameters documented in
\code{\link{pointsTile}}.
}
\author{
Christopher Adolph \email{cadolph@u.washington.edu}
}
\seealso{
\link{tile}, \link{pointsTile}, \link{scatter}
}
\keyword{dplot}
\keyword{list}

