% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Co_expression.R
\name{Co_expression}
\alias{Co_expression}
\title{TILs marker}
\usage{
Co_expression(exp_matrix)
}
\arguments{
\item{exp_matrix}{a data matrix with rows labeling features gene symbol, and columns labeling samples,expression data should be normalized values,such as "TPM"}
}
\description{
Total TILs score and cell type enrichment scores are calculated according to Danaher et al.,2017
}
\examples{
result <- Co_expression(demo_TPM)
TILs_score <- data.frame(result[[1]],check.names = FALSE)
cell_enrich <- data.frame(result[[2]],check.names = FALSE)
}
