% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_normalization.R
\name{counts_normlization}
\alias{counts_normlization}
\title{raw reads count normalization}
\usage{
counts_normlization(
  counts_matrix,
  method = "TPM",
  gene_len = "hg19_gene_symbol_length"
)
}
\arguments{
\item{counts_matrix}{a data matrix with rows labeling features gene symbol,and columns labeling samples,expression data shoule be raw reads count values}

\item{method}{a character string provide the normalization methods, can be one of c("TPM","FPKM")}

\item{gene_len}{a character string provide the gene length data,gene type is SYMBOL,gene length database can selected from c("hg19_gene_symbol_length","hg38_gene_symbol_length")}
}
\description{
raw reads count normalization
}
\examples{
result <- counts_normlization(demo_counts,method = "FPKM",gene_len = "hg38_gene_symbol_length")
}
