% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYT_marker.R
\name{CYT_calculation}
\alias{CYT_calculation}
\title{CYT marker}
\usage{
CYT_calculation(exp_matrix)
}
\arguments{
\item{exp_matrix}{a data matrix with rows labeling features gene symbol,and columns labeling samples,expression data should be normalized values,such as "TPM"}
}
\description{
The CYT (cytolytic activity) marker is estimated using the expression of granzyme A and perforin according to Rooney et al.,2015
}
\examples{
result <- CYT_calculation(demo_TPM)
}
