% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{GEP}
\alias{GEP}
\title{GEP marker}
\usage{
GEP(exp_matrix)
}
\arguments{
\item{exp_matrix}{a data matrix with rows labeling features gene symbol,and columns labeling samples}
}
\description{
T-cell–inflamed gene-expression profile (GEP),which is a biomarker that can predict response to anti–programmed cell death 1 (PD-1) therapy across multiple tumor types according to Ayers, Mark et al.,2017
}
\examples{
result <- GEP(exp_matrix = demo_TPM)
}
