% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPRES.R
\name{IPRES}
\alias{IPRES}
\title{IPRES marker}
\usage{
IPRES(exp_matrix, normalization = TRUE)
}
\arguments{
\item{exp_matrix}{a data matrix with rows labeling features gene symbol,and columns labeling samples}

\item{normalization}{a logical indicating whether expression data are normalized, such as "TPM",default is normalized values}
}
\description{
Innately resistant tumors display a transcriptional signature (referred to as the IPRES or Innate anti-PD-1 Resistance) according to Hugo et al.,2016
}
\examples{
result <- IPRES(demo_TPM)
result <- IPRES(demo_counts,normalization = FALSE)
}
