% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_boxplot.R
\name{marker_boxplot}
\alias{marker_boxplot}
\title{box plot visualization}
\usage{
marker_boxplot(score_matrix, groupinfo, paired = FALSE, fontsize = NULL)
}
\arguments{
\item{score_matrix}{a data matrix with rows labeling features immune cell type,and columns labeling samples}

\item{groupinfo}{a matrix with two columns:'sample' column labeling sampleid,'groups' column labeling group information.
when @param paired was setted as'TRUE',@param groupinfo need to provide three columns,the added column 'patient' labeling the paired relation of samples}

\item{paired}{a logical indicating whether you want a paired wilcoxon test, which is used to compare the means of two samples when each observation in one sample can be paired with an observation in the other sample,default is "FALSE"}

\item{fontsize}{a numeric value, change the font size, default is NULL(function can provide the font size automatically)}
}
\description{
Box plots for immnume signatures score with clinical features
}
\examples{
p <- marker_boxplot(score_matrix = cell_enrich,groupinfo = demo_groupinfo,fontsize = 10,paired = FALSE)
print(p)
#paired sample
p <- marker_boxplot(score_matrix = cell_enrich,groupinfo = simulated_groupinfo_paired,paired = T)
print(p)
}
