% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_heatmap.R
\name{marker_heatmap}
\alias{marker_heatmap}
\title{heatmap visualization}
\usage{
marker_heatmap(
  score_matrix,
  groupinfo,
  scaled = "row",
  title_size = NULL,
  colsize = NULL,
  rowsize = NULL
)
}
\arguments{
\item{score_matrix}{a data matrix with rows labeling features immune cell type,and columns labeling samples}

\item{groupinfo}{a matrix with two columns:'sample' column labeling sampleid,'groups' column labeling group information}

\item{scaled}{a logical indicating whether the values would be centered and scaled in either the row direction or the column direction, or none}

\item{title_size}{a numeric value, change the size of title, default is NULL(function can provide the font size of title automatically)}

\item{colsize}{a numeric value, change the size of column labels, default is NULL(function can provide the font size of column labels automatically)}

\item{rowsize}{a numeric value, change the size of row labels, default is NULL(function can provide the font size of row labels automatically)}
}
\description{
supervised analyses of cell type score between groups
}
\examples{
p <- marker_heatmap(score_matrix = cell_enrich,groupinfo = demo_groupinfo, title_size = 8)
print(p)
}
