% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_calcu.R
\name{sig_calculate}
\alias{sig_calculate}
\title{signature calculation}
\usage{
sig_calculate(exp_matrix, db)
}
\arguments{
\item{exp_matrix}{a data matrix with rows labeling features gene symbol,and columns labeling samples,expression data should be normalized values,such as "TPM"}

\item{db}{a character string can be 'integration_signature' or 'tls_sig'
integration_signature integrate tumor microenvironment-related signature,evaluated the abundance of immune cell populations in the tumor microenvironment according to Pornpimol et al.,2017 and Ott et al.,2019

tls_sig signature can be used to calculate the score of Tertiary lymphoid structures according to Tokunaga et al.,2020 and Rita et al.,2020}
}
\description{
This function can be used to calculate immnune signature
}
\examples{
result <- sig_calculate(exp_matrix = demo_TPM,db = "tls_sig")
}
