% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA.R
\name{ssgsea}
\alias{ssgsea}
\title{ssGESA method}
\usage{
ssgsea(exp_matrix, normalization = TRUE, db)
}
\arguments{
\item{exp_matrix}{a data matrix with rows labeling features gene symbol,and columns labeling samples}

\item{normalization}{a logical indicating whether expression data are normalized, such as "TPM",default is normalized values}

\item{db}{a character string can be one of c("msigdb_C2","msigdb_C5","msigdb_C6","msigdb_C7","msigdb_H","cancerSEA","TILs_28_subpopulations")
msigdb db:https://www.gsea-msigdb.org/gsea/msigdb
cancerSEA:http://biocc.hrbmu.edu.cn/CancerSEA
TILs_28_subpopulations:pan-cancer immune cell subpopulations according to Pornpimol et al.,2017}
}
\description{
use single sample gene set enrichment analysis (ssGSEA) (Barbie et al., 2009) to identify immune cell types that are over-represented in the tumor microenvironment
}
\examples{
result <- ssgsea(demo_counts,normalization = FALSE,db = "cancerSEA")
result <- ssgsea(demo_TPM,normalization = TRUE,db = "msigdb_H")
}
