% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{t test}
\usage{
t_test(score_matrix, groupinfo, paired = FALSE)
}
\arguments{
\item{score_matrix}{a data matrix with rows labeling features immune cell type,and columns labeling samples}

\item{groupinfo}{a matrix with two columns:'sample' column labeling sampleid,'groups' column labeling group information.
when @param paired are TRUE, @param groupinfo need to provide three columns,the added column 'patient' labeling the paired relation of samples}

\item{paired}{a logical indicating whether you want a paired t-test, which is used to compare the means of two samples when each observation in one sample can be paired with an observation in the other sample,default is "FALSE"}
}
\description{
For comparison of two groups,t test can be used when your data are independent and normally distributed
}
\examples{
result <- t_test(score_matrix = cell_enrich,groupinfo = demo_groupinfo,paired = F)
#paired sample
result <- t_test(score_matrix = cell_enrich,groupinfo = simulated_groupinfo_paired,paired = T)
}
