\name{tile-internal}
\alias{adjlinebreaks}
\alias{cbindfill}
\alias{convertRtoY}
\alias{convertTtoX}
\alias{delif}
\alias{fillout.axis}
\alias{fillout.axistitle}
\alias{fillout.list}
\alias{fillout.matrix}
\alias{fillout}
\alias{fillout.title}
\alias{filloutunits}
\alias{getplotnums}
\alias{getplottraces}
\alias{getplottracesall}
\alias{gridlines}
\alias{insertrow}
\alias{lighten}
\alias{lineplotTilePlot}
\alias{lineplotTilePrep}
\alias{lineplotTileTraceDefaults}
\alias{lineplotTileTraceFillout}
\alias{linesTilePlot}
\alias{linesTilePrep}
\alias{linesTileTraceDefaults}
\alias{linesTileTraceFillout}
\alias{maxUnit}
\alias{mergelist}
\alias{nightplotTilePlot}
\alias{nightplotTilePrep}
\alias{nightplotTileTraceDefaults}
\alias{nightplotTileTraceFillout}
\alias{pointsTilePlot}
\alias{pointsTilePrep}
\alias{pointsTileTraceDefaults}
\alias{pointsTileTraceFillout}
\alias{polygonTilePlot}
\alias{polygonTilePrep}
\alias{polygonTileTraceDefaults}
\alias{polygonTileTraceFillout}
\alias{polylinesTileTraceDefaults}
\alias{polylinesTileTraceFillout}
\alias{polylineTilePlot}
\alias{polylineTilePrep}
\alias{posquad}
\alias{pretty.axis}
\alias{pretty.logaxis}
\alias{rbindfill}
\alias{rgbscale}
\alias{riceplotTilePlot}
\alias{riceplotTilePrep}
\alias{riceplotTileTraceDefaults}
\alias{riceplotTileTraceFillout}
\alias{ropeladderTileAxisTitle}
\alias{ropeladderTileHarmonize}
\alias{ropeladderTilePlot}
\alias{ropeladderTilePrep}
\alias{ropeladderTileTraceDefaults}
\alias{ropeladderTileTraceFillout}
\alias{roundnums}
\alias{rugTilePlot}
\alias{rugTilePrep}
\alias{rugTileTraceDefaults}
\alias{rugTileTraceFillout}
\alias{scatterTilePlot}
\alias{scatterTilePrep}
\alias{scatterTileTraceDefaults}
\alias{scatterTileTraceFillout}
\alias{selif}
\alias{sortmc}
\alias{stacked}
\alias{switch.variables}
\alias{textTilePlot}
\alias{textTilePrep}
\alias{textTileTraceDefaults}
\alias{textTileTraceFillout}
\alias{tileAxis}
\alias{tileAxistitle}
\alias{tileAxistitleprep}
\alias{tileTitle}
\alias{tileCartouche}
\alias{tileColorcode}
\alias{tileColorspaces}
\alias{tiledimensions}
\alias{tiledimensionsIter}
\alias{tileDrawChildren}
\alias{tileDrawGrob}
\alias{tilefilloutinputs}
\alias{tilefindplotarea}
\alias{tilefitlinear}
\alias{tilefitloess}
\alias{tilefitmmest}
\alias{tilefitrobust}
\alias{tilefitwls}
\alias{tileGetBlocks}
\alias{tilegrobNest}
\alias{tileHarmonizeAxes}
\alias{tileInitWH}
\alias{tileMakePlot}
\alias{tileopendevice}
\alias{tileProcessTraces}
\alias{tilesetdefaults}
\alias{tileSetgpar}
\alias{tileSetLayer}
\alias{tilesimulates}
\alias{tileSplitBlocks}
\alias{tileWhatIf}
\alias{titlebox}
\alias{updateRxC}


\title{Internal Functions in the Tile Package}
\description{Internal functions in the tile package.}

\author{Christopher Adolph <\email{cadolph@u.washington.edu}>}
\keyword{internal}
