% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_test.R
\name{wilcox}
\alias{wilcox}
\title{wilcoxon test}
\usage{
wilcox(score_matrix, groupinfo, paired = FALSE)
}
\arguments{
\item{score_matrix}{a data matrix with rows labeling features immune cell type,and columns labeling samples.}

\item{groupinfo}{a matrix with two columns:'sample' column labeling sampleid,'groups' column labeling group information.
when @param paired was setted as'TRUE',@param groupinfo need to provide three columns,the added column 'patient' labeling the paired relation of samples}

\item{paired}{a logical indicating whether you want a paired wilcoxon test, is a non-parametric alternative to paired t-test used to compare paired data. It's used when your data are not normally distributed,default is "FALSE"}
}
\description{
For comparison of two groups,non-parametric two-sided Wilcoxon-rank sum test was used
}
\examples{
result <- wilcox(score_matrix = cell_enrich,groupinfo = demo_groupinfo,paired = F)
#paired sample
result <- wilcox(score_matrix = cell_enrich,groupinfo = simulated_groupinfo_paired,paired = T)
}
